api/requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
google-cloud-storage==2.7.0
requests==2.28.1
